requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
boto3==1.39.4
botocore==1.39.4
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
iniconfig==2.1.0
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
